Create component for current display name
